modify

+++ model/common.go
@@ -6,44 +6,44 @@ type Response struct {
 }
 
 type Video struct {
-	Id            int64  `json:"id,omitempty"`
+	Id            int32  `json:"id,omitempty"`
 	Author        User   `json:"author"`
 	PlayUrl       string `json:"play_url" json:"play_url,omitempty"`
 	CoverUrl      string `json:"cover_url,omitempty"`
-	FavoriteCount int64  `json:"favorite_count,omitempty"`
-	CommentCount  int64  `json:"comment_count,omitempty"`
+	FavoriteCount int32  `json:"favorite_count,omitempty"`
+	CommentCount  int32  `json:"comment_count,omitempty"`
 	IsFavorite    bool   `json:"is_favorite,omitempty"`
 	Title         string `json:"title,omitempty"`
 }
 
 type Comment struct {
-	Id         int64  `json:"id,omitempty"`
+	Id         int32  `json:"id,omitempty"`
 	User       User   `json:"user"`
 	Content    string `json:"content,omitempty"`
 	CreateDate string `json:"create_date,omitempty"`
 }
 
 type User struct {
-	Id            int64  `json:"id,omitempty" gorm:"id"`                        //用户id
+	Id            int32  `json:"id,omitempty" gorm:"id"`                        //用户id
 	Username      string `json:"name,omitempty" gorm:"name"`                //用户名称
-	FollowCount   int64  `json:"follow_count,omitempty" gorm:"followcount"`     //关注总数
-	FollowerCount int64  `json:"follower_count,omitempty" gorm:"followercount"` //粉丝总数
+	FollowCount   int32  `json:"follow_count,omitempty" gorm:"followcount"`     //关注总数
+	FollowerCount int32  `json:"follower_count,omitempty" gorm:"followercount"` //粉丝总数
 	IsFollow      bool   `json:"is_follow,omitempty" gorm:"isfollow"`           //true--已关注，false--未关注
 }
 
 type Message struct {
-	Id         int64  `json:"id,omitempty"`
+	Id         int32  `json:"id,omitempty"`
 	Content    string `json:"content,omitempty"`
 	CreateTime string `json:"create_time,omitempty"`
 }
 
 type MessageSendEvent struct {
-	UserId     int64  `json:"user_id,omitempty"`
-	ToUserId   int64  `json:"to_user_id,omitempty"`
+	UserId     int32  `json:"user_id,omitempty"`
+	ToUserId   int32  `json:"to_user_id,omitempty"`
 	MsgContent string `json:"msg_content,omitempty"`
 }
 
 type MessagePushEvent struct {
-	FromUserId int64  `json:"user_id,omitempty"`
+	FromUserId int32  `json:"user_id,omitempty"`
 	MsgContent string `json:"msg_content,omitempty"`
 }
